@@ -1,6 +1,6 @@
 {
   "name": "differencify",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "Perceptual diffing tool",
   "main": "dist/index.js",
   "scripts": {
@@ -8,7 +8,8 @@
     "build": "babel src --source-maps --out-dir dist --ignore '**/*.test.js'",
     "build-docs": "mkdir -p docs && jsdoc2md src/*.js > docs/api.md",
     "pretest": "npm run lint",
-    "test": "jest"
+    "test": "jest --testPathIgnorePatterns integration.test",
+    "test:integration": "jest --testPathPattern integration.test"
   },
   "repository": {
     "type": "git",
@@ -49,7 +50,7 @@
     "check-types": "^7.2.0",
     "fs": "0.0.1-security",
     "jimp": "^0.2.28",
-    "puppeteer": "^0.12.0",
+    "puppeteer": "^0.13.0",
     "type-detect": "^4.0.3"
   },
   "jest": {
@@ -62,6 +63,6 @@
     ]
   },
   "engines": {
-    "node": ">=8.6"
+    "node": ">=8.9"
   }
 }
