@@ -1,29 +1,109 @@
-const chainProxy = require('./proxyChain');
+import { isFunc, handleAsyncFunc } from './functions';
+import chain from './proxyChain';
 
-class ForTest {
+class Keyboard {
   constructor() {
-    this._value = '';
+    this._value = 'keyboard';
   }
   value() {
     return this._value;
   }
+}
+
+class Mouse {
+  constructor() {
+    this._value = 'mouse';
+  }
+  async value() {
+    return this._value;
+  }
+}
+
+class Page {
+  constructor() {
+    this.keyboard = new Keyboard();
+    this.mouse = new Mouse();
+  }
+
+  // eslint-disable-next-line class-methods-use-this
+  value() {
+    return 'Page';
+  }
+
+  async pageValue() {
+    return await this.value();
+  }
+}
+
+class Fake {
+  constructor(value) {
+    this._value = value;
+    this.page = new Page();
+  }
+  async bar() {
+    const newValue = this._value += '4';
+    return newValue;
+  }
+}
+
+class ForTest {
+  constructor() {
+    this._value = '';
+    this.page = new Page();
+  }
+
+  value() {
+    return this._value;
+  }
+
+  async pageValue() {
+    return await this.page.pageValue();
+  }
 
   _evaluateResult() { return this._value; }
-  toMatchSnapshot() { this._value = 'Executed first'; }
+  // eslint-disable-next-line class-methods-use-this
+  async _handleContinueFunc(target, property, args) {
+    return isFunc(target[property]) ? await target[property](...args) : await target[property];
+  }
+
+  _handleFunc(target, property, args) {
+    if (target === 'page') {
+      if (this[property]) {
+        return isFunc(this[property]) ? handleAsyncFunc(this, property, args) : this[property];
+      }
+      return isFunc(this.page[property]) ? handleAsyncFunc(this.page, property, args) : this.page[property];
+    }
+    return isFunc(this.page[target][property])
+      ? handleAsyncFunc(this.page[target], property, args)
+      : this.page[target][property];
+  }
 
   async echo(text) {
     this.dummy = null;
     return text;
   }
-  async f1() {
-    this._value += '1';
+
+  getValue(number) { this._value += number; return this._value; }
+
+  f1() {
+    return this.getValue('1');
   }
-  async f2() {
-    this._value += '2';
+  f2() {
+    return this.getValue('2');
   }
-  async f3() {
-    this._value += '3';
+  f3() {
+    return this.getValue('3');
   }
+
+  f4() {
+    const fake = new Fake(this._value);
+    return fake;
+  }
+
+  withArgs(args) {
+    return this.getValue(args);
+  }
+
   async throwError() {
     this.dummy = null;
     throw new Error('error');
@@ -32,7 +112,7 @@ class ForTest {
 
 describe('chainProxy', () => {
   it('add chain interface', async () => {
-    const obj = chainProxy(new ForTest());
+    const obj = chain(new ForTest());
     await obj
       .f1()
       .f2()
@@ -44,7 +124,7 @@ describe('chainProxy', () => {
       });
   });
   it('add chain interface.(reverse)', async () => {
-    const obj = chainProxy(new ForTest());
+    const obj = chain(new ForTest());
     await obj
       .f3()
       .f2()
@@ -55,61 +135,8 @@ describe('chainProxy', () => {
         expect(value).toEqual('321');
       });
   });
-  it('executes toMatchSnapshot first', async () => {
-    const obj = chainProxy(new ForTest());
-    await obj
-      .f1()
-      .toMatchSnapshot()
-      .value()
-      .end()
-      .then((value) => {
-        expect(value).toEqual('Executed first1');
-      });
-  });
-  it('remove duplicate toMatchSnapshot steps', async () => {
-    const obj = chainProxy(new ForTest());
-    await obj
-      .f1()
-      .toMatchSnapshot()
-      .toMatchSnapshot()
-      .value()
-      .end()
-      .then((value) => {
-        expect(value).toEqual('Executed first1');
-      });
-  });
-  describe('#result', () => {
-    it('receive last return value', async () => {
-      const obj = chainProxy(new ForTest());
-      let result = null;
-      await obj
-        .echo('AAA')
-        .result((v) => { result = v; })
-        .end()
-        .then(() => {
-          expect(result).toEqual('AAA');
-        });
-    });
-  });
-  it('result function', async () => {
-    const obj = chainProxy(new ForTest());
-    let result = null;
-    await obj
-      .echo('AAA')
-      .result((v) => { result = v; })
-      .end()
-      .then(() => {
-        expect(result).toEqual('AAA');
-      });
-  });
-  it('returns target', () => {
-    const target = new ForTest();
-    target.v = 100;
-    const obj = chainProxy(target);
-    expect(obj.target.v).toEqual(100);
-  });
   it('removes all actions if occurred an exception.', async () => {
-    const obj = chainProxy(new ForTest());
+    const obj = chain(new ForTest());
     await obj
       .f1()
       .throwError()
@@ -120,13 +147,150 @@ describe('chainProxy', () => {
       })
       .catch(() => {
         obj
+        .f3()
         .end()
-        .then(() => {
-          expect(obj.target.value()).toEqual('1');
+        .then((result) => {
+          expect(result).toEqual('13');
         })
         .catch((e) => {
           throw (e);
         });
       });
+  });
+  describe('#result', () => {
+    it('receive last return value', async () => {
+      const obj = chain(new ForTest());
+      let result = null;
+      await obj
+        .echo('AAA')
+        .result((v) => { result = v; })
+        .end()
+        .then(() => {
+          expect(result).toEqual('AAA');
+        });
+    });
+    it('result function', async () => {
+      const obj = chain(new ForTest());
+      let result = null;
+      await obj
+        .echo('AAA')
+        .result((v) => { result = v; })
+        .end()
+        .then(() => {
+          expect(result).toEqual('AAA');
+        });
+    });
+  });
+  describe('#then', () => {
+    it('continues to current target.', async () => {
+      const target = new ForTest();
+      const obj = chain(target);
+      target.chainedTarget = obj;
+      await obj
+        .f1()
+        .f4()
+          .then
+          .bar()
+        .end()
+        .then((result) => {
+          expect(result).toEqual('14');
+        });
+    });
+    it('will switch to main target after continues to current target.', async () => {
+      const target = new ForTest();
+      const obj = chain(target);
+      target.chainedTarget = obj;
+      await obj
+        .f1()
+        .f4()
+          .then
+          .bar()
+        .f2()
+        .end()
+        .then((result) => {
+          expect(result).toEqual('12');
+        });
+    });
+  });
+  describe('subModules', () => {
+    it('continues to current target.', async () => {
+      const target = new ForTest();
+      const obj = chain(target);
+      target.chainedTarget = obj;
+      await obj
+        .f1()
+        .f4()
+        .keyboard
+          .value()
+        .end()
+        .then((result) => {
+          expect(result).toEqual('keyboard');
+        });
+    });
+    it('continues to another target.', async () => {
+      const target = new ForTest();
+      const obj = chain(target);
+      target.chainedTarget = obj;
+      await obj
+        .f1()
+        .keyboard
+          .value()
+        .page
+          .f4()
+            .then
+            .bar()
+          .mouse
+            .value()
+        .end()
+        .then((result) => {
+          expect(result).toEqual('mouse');
+        });
+    });
+    it('continues to another target and get result.', async () => {
+      const target = new ForTest();
+      const obj = chain(target);
+      target.chainedTarget = obj;
+      await obj
+        .f1()
+        .keyboard
+          .value()
+        .page
+          .f4()
+            .then
+            .bar()
+            .result((result) => { expect(result).toEqual('14'); })
+          .mouse
+            .value()
+        .end()
+        .then((result) => {
+          expect(result).toEqual('mouse');
+        });
+    });
+  });
+  describe('unchained', () => {
+    it('will work unchained', async () => {
+      const obj = chain(new ForTest(), false);
+      const value = await obj.f1();
+      expect(value).toEqual('1');
+    });
+    it('will work with complex scenarios', async () => {
+      const obj = chain(new ForTest(), false);
+      const value = await obj.f1();
+      const keyboardValue = await obj.keyboard.value();
+      expect(value).toEqual('1');
+      expect(keyboardValue).toEqual('keyboard');
+    });
+    it('will work with calling async function', async () => {
+      const obj = chain(new ForTest(), false);
+      const f1Value = await obj.f1();
+      const pageValue = await obj.pageValue();
+      expect(f1Value).toEqual('1');
+      expect(pageValue).toEqual('Page');
+    });
+    it('will pass the right arguments', async () => {
+      const obj = chain(new ForTest(), false);
+      const withArgsValue = await obj.withArgs('param');
+      expect(withArgsValue).toEqual('param');
+    });
   });
 });
