@@ -1,35 +1,43 @@
 /* eslint-disable prefer-rest-params */
 import logger from '../utils/logger';
-
-const RESULT_FUNCTION_NAME = '___RESULT_FUNCTION';
+import moduleList from './moduleList';
 
 class ChainObject {
   constructor(target, options) {
     this.target = target;
     this.options = options;
+    this.currentTarget = 'page';
     this.actions = [];
   }
 
-  addAction(name, args) {
-    this.actions.push({ name, args });
+  setCurrentTarget(currentTarget) {
+    this.currentTarget = currentTarget;
+  }
+
+  addAction({
+      target = this.currentTarget,
+      property = undefined,
+      args = undefined,
+    } = {}) {
+    this.actions.push({ target, property, args });
   }
 
   async end() {
     let result = null;
-    let actions = this.actions;
+    let continueChain = false;
+    const actions = this.actions;
     this.actions = [];
-    actions.forEach((action) => {
-      if (action.name === 'toMatchSnapshot') {
-        this.target[action.name](...action.args);
-      }
-    });
-    actions = actions.filter(action => action.name !== 'toMatchSnapshot');
     // eslint-disable-next-line no-restricted-syntax
     for (const action of actions) {
-      if (RESULT_FUNCTION_NAME === action.name) {
+      if (continueChain) {
+        result = await this.target._handleContinueFunc(result, action.property, action.args);
+        continueChain = false;
+      } else if (action.property === this.options.continueFunc) {
+        continueChain = true;
+      } else if (action.property === this.options.resultFunc) {
         result = await action.args[0].apply(null, [result]);
       } else {
-        result = await this.target[action.name](...action.args);
+        result = await this.target[this.options.funcHandler](action.target, action.property, action.args);
       }
     }
     return result;
@@ -38,44 +46,48 @@ class ChainObject {
 
 const makeHandler = (target, options) =>
   ({
-    get: (chainObj, name) => {
-      if (name === options.endFuncName) {
-        return () => {
-          chainObj.addAction(options.updateFunctionName, arguments);
-          return chainObj.end(chainObj, arguments)
+    get: (chainObj, property) => {
+      if (!options.chained) {
+        return (moduleList.includes(property))
+            ? target[options.funcHandler]('page', property)
+            : function handle() {
+              return target[options.funcHandler]('page', property, arguments);
+            };
+      }
+      if (property === options.endFunc) {
+        return () =>
+          chainObj.end()
             .then(result => result)
             .catch((e) => {
-              logger.error(e);
+              logger.trace(e);
               throw e;
             });
-        };
-      } else if (name === options.resultFuncName) {
-        return function handle() {
-          chainObj.addAction(RESULT_FUNCTION_NAME, arguments);
-          return this;
-        };
-      } else if (name === 'target') {
-        return target;
-      } else if (typeof (target[name]) === 'function') {
-        return function handle() {
-          chainObj.addAction(name, arguments);
-          return this;
-        };
-      } else if (name in target) {
-        return target[name];
       }
-      throw new Error(`'${name}' is not defined on a target object.`);
+      if (moduleList.includes(property)) {
+        chainObj.setCurrentTarget(property);
+        return target.chainedTarget;
+      }
+      if (property === options.continueFunc) {
+        chainObj.addAction({ property });
+        return target.chainedTarget;
+      }
+      return function handle() {
+        chainObj.addAction({ property, args: arguments });
+        return this;
+      };
     },
     set: (chainObj, name, value) => {
       throw new Error(`You cannot set a ${value} to Proxy object.`);
     },
   });
 
-const chainProxy = (target) => {
+const chain = (target, chained = true) => {
   const defaultParams = {
-    resultFuncName: 'result',
-    endFuncName: 'end',
-    updateFunctionName: '_evaluateResult',
+    resultFunc: 'result',
+    endFunc: 'end',
+    continueFunc: 'then',
+    funcHandler: '_handleFunc',
+    chained,
   };
   const chainObject = new ChainObject(target, defaultParams);
   return new Proxy(
@@ -84,4 +96,4 @@ const chainProxy = (target) => {
   );
 };
 
-module.exports = chainProxy;
+module.exports = chain;
