@@ -1,7 +1,7 @@
 import puppeteer from 'puppeteer';
-import chainProxy from './helpers/proxyChain';
+import chain from './helpers/proxyChain';
 import { sanitiseGlobalConfiguration, sanitiseTestConfiguration } from './sanitiser';
-import Page from './page';
+import Target from './target';
 import logger from './utils/logger';
 
 export default class Differencify {
@@ -14,17 +14,46 @@ export default class Differencify {
     this.testId = 0;
   }
 
-  async launchBrowser() {
+  async launchBrowser(options) {
     if (!this.browser) {
       logger.log('Launching browser...');
       try {
-        this.browser = await puppeteer.launch(this.configuration.puppeteer);
+        this.browser = await puppeteer.launch(options);
       } catch (error) {
-        logger.error(error);
+        logger.trace(error);
       }
     } else {
       logger.log('Using existing browser instance');
     }
+  }
+
+  static executablePath() {
+    return puppeteer.executablePath();
+  }
+
+  static chromeExecutablePath() {
+    return puppeteer.executablePath();
+  }
+
+  async launch(options) {
+    this.launchBrowser(options);
+  }
+
+  async connectBrowser(options) {
+    if (!this.browser) {
+      logger.log('Launching browser...');
+      try {
+        this.browser = await puppeteer.connect(options);
+      } catch (error) {
+        logger.trace(error);
+      }
+    } else {
+      logger.log('Using existing browser instance');
+    }
+  }
+
+  async connect(options) {
+    this.connectBrowser(options);
   }
 
   init(config) {
@@ -33,11 +62,11 @@ export default class Differencify {
     if (testConfig.isUpdate) {
       logger.warn('Your tests are running on update mode. Test screenshots will be updated');
     }
-    const page = new Page(this.browser, testConfig, this.configuration);
-    if (testConfig.chain) {
-      return chainProxy(page);
-    }
-    return page;
+    const target = new Target(this.browser, this.configuration, testConfig);
+    target.isJest();
+    const chainedTarget = chain(target, testConfig.chain);
+    target.chainedTarget = chainedTarget;
+    return chainedTarget;
   }
 
   async cleanup() {
@@ -47,7 +76,7 @@ export default class Differencify {
         await this.browser.close();
       }
     } catch (error) {
-      logger.error(error);
+      logger.trace(error);
     }
   }
 }
