@@ -2,9 +2,9 @@ import puppeteer from 'puppeteer';
 import chainProxy from './helpers/proxyChain';
 import Differencify from './index';
 import logger from './utils/logger';
-import Page from './page';
+import Target from './target';
 
-jest.mock('./page');
+jest.mock('./target');
 jest.mock('./helpers/proxyChain');
 
 jest.mock('puppeteer', () => ({
@@ -34,8 +34,7 @@ describe('Differencify', () => {
     chainProxy.mockClear();
   });
   it('launchBrowser', async () => {
-    await differencify.launchBrowser();
-    expect(puppeteer.launch).toHaveBeenCalledWith({
+    const browserOptions = {
       args: [],
       dumpio: false,
       executablePath: undefined,
@@ -43,7 +42,22 @@ describe('Differencify', () => {
       ignoreHTTPSErrors: false,
       slowMo: 0,
       timeout: 30000,
-    });
+    };
+    await differencify.launchBrowser(browserOptions);
+    expect(puppeteer.launch).toHaveBeenCalledWith(browserOptions);
+    await differencify.launch(browserOptions);
+    expect(puppeteer.launch).toHaveBeenCalledWith(browserOptions);
+    expect(logger.log).toHaveBeenCalledWith('Launching browser...');
+  });
+  it('connect', async () => {
+    const browserOptions = {
+      browserWSEndpoint: 'endpoint',
+      ignoreHTTPSErrors: false,
+    };
+    await differencify.launchBrowser(browserOptions);
+    expect(puppeteer.launch).toHaveBeenCalledWith(browserOptions);
+    await differencify.launch(browserOptions);
+    expect(puppeteer.launch).toHaveBeenCalledWith(browserOptions);
     expect(logger.log).toHaveBeenCalledWith('Launching browser...');
   });
   it('does not relaunch browser if one browser instance exists', async () => {
@@ -59,29 +73,20 @@ describe('Differencify', () => {
   it('init without chaining', async () => {
     process.env.update = true;
     await differencify.init({ chain: false });
-    expect(Page).toHaveBeenCalledWith(null,
-      {
-        chain: false,
-        newWindow: false,
-        testName: 'test1',
-        isUpdate: 'true',
-      },
+    expect(Target).toHaveBeenCalledWith(null,
       {
         debug: false,
         mismatchThreshold: 0.001,
-        puppeteer: {
-          args: [],
-          dumpio: false,
-          executablePath: undefined,
-          headless: true,
-          ignoreHTTPSErrors: false,
-          slowMo: 0,
-          timeout: 30000,
-        },
         saveDifferencifiedImage: true,
         imageSnapshotPath: 'differencify_reports',
+      },
+      {
+        chain: false,
+        testName: 'test',
+        isUpdate: 'true',
+        testId: 1,
       });
-    expect(chainProxy).toHaveBeenCalledTimes(0);
+    expect(chainProxy).toHaveBeenCalledTimes(1);
     expect(logger.warn).toHaveBeenCalledWith('Your tests are running on update mode. Test screenshots will be updated');
     delete process.env.update;
   });
